Comment constant to strip from the top of TS files
Motivation: this comment will be included at the top of TS files written in pure JS to clarify why there may be some TS warnings. The files are TS files in order to differentiate them from the output JS files. Before producing the final output, the comment will be stripped. This constant provides the value to look for.