Refactor - React: `OneSignalFunctionTemplate` Implementation
This refactor makes slight changes to the OneSignal react function code as well as changes the argument passed into the template. Instead of passing the name and arguments individually to the function, we simply pass the entire `IFunctionSignature` object. This will help in codebases where the index file extension is .ts as we will need the information regarding the individual argument types.