Add optional subdir argument to CodeGenManager
Motivation: in the case of the Vue shim, we want to create two separate builds. One for Vue2 and one for Vue3.

This optional argument will allow us to create a separate subdirectory for each.

We modify the build path to include this argument. If none is passed in, the path will simply resolve to the original top-level build directory for that shim.