Convert `writeOneSignalFunctions` to recursive function
Motivation: we give the function a "namespace chain" to start and the function will recurse down the tree based on the given api spec. As it does, it adds to the namespace chain. This chain is passed to the function template writer in order to appropriately nest the wrapped function call.