Remove function overload constants without `function` prefix
Because we changed the function names to have the namespace prefix to avoid collisions, it no longer makes sense to define the functions within the class.

Thus, we no longer need the prefix-less override since we will define the functions *outside* the class, then simply assign them to the correctly-named property (matches the API).