Support writing properties for all namespaces
Currently, we are hardcoding for the `PushSubscription` namespace only since it was the only one to have properties. But with `permissionNative` on the `Notifications` namespace, we now have to support writing properties on any namespace.

This is tricky because the namespaces have different levels of nesting.

To address this, we define a PARENT_PATH_FOR_NAMESPACE_MAP that gives the string for the namespace path from `OneSignal` (root) to the target

Also, defines a PROPERTY_DEFAULT_MAP to set the default value for `permissionNative`