More robust window check
Fixes https://github.com/OneSignal/react-onesignal/issues/119

This error is typical when you're trying to reference the window object in a server-side context. In a Next.js application for example, code can run both on the server and the client. The window object is specific to the browser and doesn't exist on the server. Hence, when you try to access window during the server-side rendering phase, you get an error.